[modified]アバターデータダウンロード時の非同期処理をUniTaskに移行
